adds delete, edit user functionality f:; cd 'f:\Projects\JavaProjects\Shop_Project'; & 'C:\Users\Micro\.vscode\extensions\redhat.java-1.13.0-win32-x64\jre\17.0.5-win32-x86_64\bin\java.exe' '@C:\Users\Micro\AppData\Local\Temp\cp_5qz3d3s63ev8msxch8o23x9t3.argfile' 'com.shop.admin.ShopBackEndApplication'

+++ Shop_WebParent/ShopBackEnd/src/main/java/com/shop/admin/service/UserService.java
@@ -1,5 +1,6 @@
 package com.shop.admin.service;
 
+import com.shop.admin.exception.UserNotFoundException;
 import com.shop.admin.repository.RoleRepository;
 import com.shop.admin.repository.UserRepository;
 import com.shop.model.Role;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -26,29 +28,56 @@ public class UserService {
         return userRepository.findAll();
     }
 
+    public User findById(Long id) throws UserNotFoundException {
+        Optional<User> found = userRepository.findById(id);
+        if (found.isEmpty())
+            throw new UserNotFoundException("Could not find the user with id " + id);
+        return found.get();
+    }
+
     public List<Role> findAllRoles() {
         return roleRepository.findAll();
     }
 
-    public User getByEmail(String email) {
+    public User getByEmail(String email) throws UserNotFoundException {
         Optional<User> user = userRepository.findByEmail(email);
         if (user.isEmpty())
-            throw new IllegalArgumentException("User doesn't exist");
+            throw new UserNotFoundException("User doesn't exist");
         return user.get();
     }
 
-    public boolean isEmailExist(String email) {
-        return userRepository.existsByEmail(email);
+    public boolean isEmailUnique(Long id, String email) {
+        Optional<User> user = userRepository.findByEmail(email);
+        if (user.isEmpty())
+            return true;
+        else if (id == null)
+            return false;
+        else if (!Objects.equals(user.get().getId(), id))
+            return false;
+
+        return true;
     }
 
     @Transactional
     public void save(User user) {
-        user.setPassword(encoder.encode(user.getPassword()));
+        if (user.getId() != null) {
+            Optional<User> optional = userRepository.findById(user.getId());
+            if (optional.isPresent()) {
+                User alrdyExist = optional.get();
+                if (alrdyExist.getPassword().isEmpty())
+                    user.setPassword(alrdyExist.getPassword());
+                else
+                    user.setPassword(encoder.encode(user.getPassword()));
+            }
+        } else user.setPassword(encoder.encode(user.getPassword()));
         userRepository.save(user);
     }
 
     @Transactional
-    public void delete(long id) {
+    public void delete(Long id) throws UserNotFoundException {
+        Long counted = userRepository.countById(id);
+        if (counted == null || counted == 0)
+            throw new UserNotFoundException("Could not find user with this ID " + id);
         userRepository.deleteById(id);
     }
 }
